Added integration tests
