+++ tests/test_misc.rs
@@ -1,5 +1,20 @@
-// use lazerpay::{self, config::ApiConfig};
-// use dotenv::dotenv;
-// use reqwest::Client;
+mod common;
+
+use common::{set_up, TestResult};
+use lazerpay::{misc::*};
+
+
+
+#[tokio::test]
+async fn test() -> TestResult {
+    let (client, config) = set_up();
+
+    let misc_client = Misc::new(&config, &client);
+
+    let _resp = misc_client.get_accepted_coins().await?;
+    let _resp2 = misc_client.get_balance("USDT").await?;
+    let _resp3 = misc_client.get_rate("USDT", "USD").await?;
+    Ok(())
+}
